Edit page functional with auth, responsive images
